+++ apps/games/components/ImageSearchSelector/Results.tsx
@@ -1,25 +1,28 @@
 "use client"
-import type { Image } from "@phylopic/api-models"
+import type { ImageWithEmbedded } from "@phylopic/api-models"
 import { SearchContext } from "@phylopic/client-components"
 import { ImageThumbnailView } from "@phylopic/ui"
 import { useContext, type FC } from "react"
 import styles from "./Results.module.scss"
+import { UUID } from "@phylopic/utils"
 export interface Props {
-    onSelect: (value: Image | undefined) => void
+    excluded?: ReadonlySet<UUID>
+    onSelect: (value: ImageWithEmbedded | undefined) => void
 }
-export const Results: FC<Props> = ({ onSelect }) => {
+export const Results: FC<Props> = ({ excluded, onSelect }) => {
     const [state] = useContext(SearchContext) ?? []
-    console.debug(state)
     const results = state?.imageResults ?? []
     return (
         <ul className={styles.main}>
-            {results.map(result => (
-                <li key={result.uuid} className={styles.item}>
-                    <button onClick={() => onSelect(result)} className={styles.button}>
-                        <ImageThumbnailView value={result} />
-                    </button>
-                </li>
-            ))}
+            {results
+                .filter(result => !excluded?.has(result.uuid))
+                .map(result => (
+                    <li key={result.uuid} className={styles.item}>
+                        <button onClick={() => onSelect(result)} className={styles.button}>
+                            <ImageThumbnailView value={result} />
+                        </button>
+                    </li>
+                ))}
         </ul>
     )
 }